+++ source/libphonenumber/RFC3966.js
@@ -1,0 +1,1 @@
+export { parseRFC3966, formatRFC3966 } from 'libphonenumber-js/core'