+++ source/libphonenumber/parsePhoneNumber.js
@@ -1,0 +1,1 @@
+export { parsePhoneNumberFromString as default } from 'libphonenumber-js/core'