+++ source/libphonenumber/parsePhoneNumberDefaultMetadata.test.js
@@ -1,0 +1,7 @@
+import parsePhoneNumber from './parsePhoneNumberDefaultMetadata'
+
+describe('parsePhoneNumber (default metadata)', () => {
+	it('should parse phone numbers', () => {
+		parsePhoneNumber('+12133734253').country.should.equal('US')
+	})
+})