+++ test/exports.test.js
@@ -1,6 +1,9 @@
 import PhoneInputDefault,
 {
 	PhoneInput,
+	parseRFC3966,
+	formatRFC3966,
+	parsePhoneNumber,
 	formatPhoneNumber,
 	formatPhoneNumberIntl,
 	isValidPhoneNumber
@@ -13,6 +16,9 @@ describe(`exports`, function()
 	{
 		PhoneInputDefault.should.be.a('function')
 		PhoneInput.should.be.a('function')
+		parseRFC3966('tel:+12133734253;ext=123').ext.should.equal('123')
+		formatRFC3966({ number: '+12133734253', ext: '123' }).should.equal('tel:+12133734253;ext=123')
+		parsePhoneNumber('+78005553535').country.should.equal('RU')
 		formatPhoneNumber('+78005553535').should.equal('8 (800) 555-35-35')
 		formatPhoneNumberIntl('+78005553535').should.equal('+7 800 555 35 35')
 		isValidPhoneNumber('+78005553535').should.equal(true)
@@ -24,6 +30,9 @@ describe(`exports`, function()
 		const Custom = require('../custom')
 
 		Library.default.should.be.a('function')
+		Library.parseRFC3966('tel:+12133734253;ext=123').ext.should.equal('123')
+		Library.formatRFC3966({ number: '+12133734253', ext: '123' }).should.equal('tel:+12133734253;ext=123')
+		Library.parsePhoneNumber('+78005553535').country.should.equal('RU')
 		Library.formatPhoneNumber('+78005553535').should.equal('8 (800) 555-35-35')
 		Library.formatPhoneNumberIntl('+78005553535').should.equal('+7 800 555 35 35')
 		Library.isValidPhoneNumber('+78005553535').should.equal(true)